Refactor and change CLI to use cobra
