+++ .gitignore
@@ -1,1 +1,1 @@
-./kubeclean
+kubeclean
