+++ go.mod
@@ -4,7 +4,10 @@ go 1.15
 
 require (
 	github.com/manifoldco/promptui v0.8.0
-	k8s.io/api v0.19.0
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/spf13/cobra v1.1.3
+	github.com/spf13/viper v1.7.1
+	k8s.io/api v0.19.0 // indirect
 	k8s.io/apimachinery v0.19.0
 	k8s.io/client-go v0.19.0
 )
